Rename data attribute to conform to standard
Dashes are supposed to be used in HTML, rather than uppercase letters. The browser automatically converts the name to camelCase in the `dataset` property.

The `(int)` cast was removed because it's redundant; `bbp_get_displayed_user_id()` always returns an `int`.

+++ settings/settings.php
@@ -46,7 +46,7 @@ function render_custom_ui() : void {
 
 	wp_enqueue_style( 'wp-components' );
 
-	$user_id    = (int) bbp_get_displayed_user_id();
+	$user_id    = bbp_get_displayed_user_id();
 	$json_attrs = json_encode( [ 'userId' => $user_id ] );
 
 	$preload_paths = [
